Update 24-09-2022: General clean & added generating scripts
Old files that are no longer in use are now removed.

Two scripts are now available to generate webpage content.
sort_blog_index.py which is a python script that generates a list of
paths to blog files in order depending on the metadata cointained in
them refered as "article-date", this script in a future may be modified
to sort the blog files depending on the other metadata values like "article-name",
as well as sort the paths names in reverse order, etc. The other script
added sync.sh is more general, and makes use of the script names before "sort_blog_index.py"
to generate the "blog_index.shtml" file, as well as an rss feed file
(rss.xml).