Added CC License footer & reduced fonts size/loading times
In order to reduce the fonts loading times, in particular the font
'Futura Medium', which is the one used in the header title, before
loading it from the server the web browser tries to find it on local
storage.

Now the footer shows that the page is under CC-BY 4.0 license and an
image is shown.

The awesome-fonts package used previously gets deprecated in favor of
a heavly strip down version of it (contains only the icon used).