syntax-highlight script (binary), md folder, webp exclude
Added a syntax-highlight script written in go which is not working for
the moment, but the css stylesheet is already modified to work with
these script. The part that is not working properly is when it's
executed from `./build_page.sh` and with a file that is not
`dwm-config.md`.

webp_exclude file added which is meant to be copied to the server so it
knows which files should not copy or sync.

article-title-with-icon.html is a template to use when a md file
contains an article title logo, so it changes the hole title and inserts
a div with the logo and the title, so it is easier to apply style with
css.