add new sections `files`
Adds new section in the top links bar, this new sections is a directory
listing of a new folder `/files` located in the root of the
project. The folder files are previewed with nginx built-in directory
listing and to make it nicer three files are included in `/css` folder,
these files are then included after and before the body of the generated
webpage by nginx, to make it work requires some configuration in the
nginx configuration file.